refactor: move navigation bar tests to separate file
